Apply a function to elements of cytoset, GatingHierarchy or GatingSet

cyto_apply is convenient wrapper around lapply and apply to apply a function over elements of a cytoset, GatingHierarchy or GatingSet. cyto_apply is extremely flexible by supporting functions that accept the data in either cytoset, cytoframe, matrix or vector formats. All the data processing steps are handled internally by cyto_data_extract prior to passing the data to the specified function. It is important that the arguments in FUN do not conflict with the arguments of cyto_apply and they should be supplied to cyto_apply by name.

# S3 method for list
cyto_apply(
  x,
  FUN,
  ...,
  simplify = TRUE,
  parent = "root",
  input = "cytoframe",
  copy = TRUE,
  channels = NULL,
  trans = NA,
  inverse = FALSE
)

Arguments

x
object of class cytoset, GatingHierarchy, GatingSet.

FUN
name of a function to apply to each element of x.

...
additional arguments passed to FUN. Multiple arguments are supported but must be named, see examples below.

simplify
logical indicating whether attempts should be made to coerce the output to a cytoset or matrix, set to TRUE by default.

parent
name of the parent(s) population to extract from GatingHierarchy or GatingSet objects, set to "root" by default.

input
indicates the data input format as required by FUN can be either 1 - "cytoset", 2 - "cytoframe", 3 - "matrix", 4 - "column" or 5 - "row", set to "cytoframe" by default. cyto_apply will take care of all the data formatting prior to passing it FUN. The "column" and "row" options are for functions that expect vectors as the input.

copy
logical indicating whether the data should be copied prior to preprocessing the data, set to TRUE by default to ensure that the original data remains unchanged.

channels
vector of channels which should be included in the data passed to FUN, set to all channels by default.

trans
object of class transformerList containing the definitions of the transformers applied to the supplied data. These transformers are only required when a cytoset is supplied to inverse transformations when inverse = TRUE.

inverse
logical indicating whether the data should be inverse transformed prior to applying FUN, set to FALSE by default.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Compute CD44 & CD69 quantiles for CD4 T Cells & CD8 T Cells
cyto_apply(gs,
"quantile",
probs = c(0.25, 0.5, 0.75),
na.rm = TRUE,
input = "column",
parent = c("CD4 T Cells", "CD8 T Cells"),
channels = c("CD44", "CD69"))

#> $`CD4 T Cells`
#>                       Alexa Fluor 647-A  7-AAD-A
#> Activation_1.fcs|25%           1.392362 1.293839
#> Activation_1.fcs|50%           1.826012 1.678628
#> Activation_1.fcs|75%           2.219541 1.952750
#> Activation_2.fcs|25%           1.367154 1.222302
#> Activation_2.fcs|50%           1.785929 1.630992
#> Activation_2.fcs|75%           2.163875 1.925536
#> Activation_3.fcs|25%           1.454165 1.137821
#> Activation_3.fcs|50%           1.836228 1.621658
#> Activation_3.fcs|75%           2.204394 1.917658
#> Activation_4.fcs|25%           1.473235 1.146265
#> Activation_4.fcs|50%           1.883880 1.603878
#> Activation_4.fcs|75%           2.238237 1.935748
#> Activation_5.fcs|25%           1.466364 1.151078
#> Activation_5.fcs|50%           1.921073 1.613258
#> Activation_5.fcs|75%           2.308435 1.928200
#> Activation_6.fcs|25%           1.477876 1.145710
#> Activation_6.fcs|50%           1.943532 1.612922
#> Activation_6.fcs|75%           2.328214 1.928318
#> Activation_7.fcs|25%           1.535726 1.197653
#> Activation_7.fcs|50%           1.956435 1.674200
#> Activation_7.fcs|75%           2.291828 1.991175
#> Activation_8.fcs|25%           1.553202 1.163052
#> Activation_8.fcs|50%           1.955033 1.648277
#> Activation_8.fcs|75%           2.319714 2.016022
#> Activation_9.fcs|25%           1.376841 1.174790
#> Activation_9.fcs|50%           1.807486 1.600126
#> Activation_9.fcs|75%           2.172575 1.921031
#> Activation_10.fcs|25%          1.365435 1.119428
#> Activation_10.fcs|50%          1.788606 1.594017
#> Activation_10.fcs|75%          2.164280 1.927163
#> Activation_11.fcs|25%          1.339373 1.122655
#> Activation_11.fcs|50%          1.803123 1.605850
#> Activation_11.fcs|75%          2.158658 1.914451
#> Activation_12.fcs|25%          1.443485 1.191767
#> Activation_12.fcs|50%          1.888334 1.641422
#> Activation_12.fcs|75%          2.232161 1.958990
#> Activation_13.fcs|25%          1.399672 1.130619
#> Activation_13.fcs|50%          1.845230 1.581985
#> Activation_13.fcs|75%          2.194477 1.918822
#> Activation_14.fcs|25%          1.489415 1.096635
#> Activation_14.fcs|50%          1.881393 1.571105
#> Activation_14.fcs|75%          2.285051 1.932304
#> Activation_15.fcs|25%          1.509318 1.144117
#> Activation_15.fcs|50%          1.930057 1.661857
#> Activation_15.fcs|75%          2.314642 2.033277
#> Activation_16.fcs|25%          1.576347 1.236173
#> Activation_16.fcs|50%          1.966340 1.715102
#> Activation_16.fcs|75%          2.352344 2.053434
#> Activation_17.fcs|25%          1.462377 1.266754
#> Activation_17.fcs|50%          1.871532 1.685941
#> Activation_17.fcs|75%          2.245615 1.975144
#> Activation_18.fcs|25%          1.347205 1.170189
#> Activation_18.fcs|50%          1.775734 1.604249
#> Activation_18.fcs|75%          2.183629 1.939864
#> Activation_19.fcs|25%          1.479578 1.184695
#> Activation_19.fcs|50%          1.894230 1.624686
#> Activation_19.fcs|75%          2.224761 1.937973
#> Activation_20.fcs|25%          1.430897 1.154355
#> Activation_20.fcs|50%          1.832697 1.594409
#> Activation_20.fcs|75%          2.163041 1.903063
#> Activation_21.fcs|25%          1.508437 1.179577
#> Activation_21.fcs|50%          1.925350 1.632671
#> Activation_21.fcs|75%          2.255044 1.967472
#> Activation_22.fcs|25%          1.462251 1.146282
#> Activation_22.fcs|50%          1.885998 1.626843
#> Activation_22.fcs|75%          2.232862 1.960761
#> Activation_23.fcs|25%          1.578691 1.376039
#> Activation_23.fcs|50%          2.010528 1.940222
#> Activation_23.fcs|75%          2.722136 2.538155
#> Activation_24.fcs|25%          1.623607 1.348271
#> Activation_24.fcs|50%          2.059333 1.873260
#> Activation_24.fcs|75%          2.626347 2.441356
#> Activation_25.fcs|25%          1.451097 1.187235
#> Activation_25.fcs|50%          1.854646 1.604211
#> Activation_25.fcs|75%          2.223547 1.930104
#> Activation_26.fcs|25%          1.341361 1.127754
#> Activation_26.fcs|50%          1.791116 1.566863
#> Activation_26.fcs|75%          2.148130 1.885356
#> Activation_27.fcs|25%          1.356639 1.177639
#> Activation_27.fcs|50%          1.775762 1.617650
#> Activation_27.fcs|75%          2.147394 1.936728
#> Activation_28.fcs|25%          1.345862 1.148808
#> Activation_28.fcs|50%          1.804809 1.584933
#> Activation_28.fcs|75%          2.235712 1.926529
#> Activation_29.fcs|25%          1.517821 1.246082
#> Activation_29.fcs|50%          1.941772 1.743630
#> Activation_29.fcs|75%          2.365905 2.168920
#> Activation_30.fcs|25%          1.450513 1.153180
#> Activation_30.fcs|50%          1.847587 1.629187
#> Activation_30.fcs|75%          2.202010 2.040919
#> Activation_31.fcs|25%          1.625555 1.464401
#> Activation_31.fcs|50%          2.100993 2.024567
#> Activation_31.fcs|75%          2.835549 2.710259
#> Activation_32.fcs|25%          1.559602 1.390927
#> Activation_32.fcs|50%          1.996726 1.914871
#> Activation_32.fcs|75%          2.725239 2.513524
#> Activation_33.fcs|25%                NA       NA
#> Activation_33.fcs|50%                NA       NA
#> Activation_33.fcs|75%                NA       NA
#> 
#> $`CD8 T Cells`
#>                       Alexa Fluor 647-A   7-AAD-A
#> Activation_1.fcs|25%          0.9941716 0.8496584
#> Activation_1.fcs|50%          1.5937119 1.2438118
#> Activation_1.fcs|75%          2.4289400 1.5948688
#> Activation_2.fcs|25%          0.9832608 0.8549758
#> Activation_2.fcs|50%          1.5441628 1.2833701
#> Activation_2.fcs|75%          2.3831291 1.6425841
#> Activation_3.fcs|25%          0.9972992 0.8119996
#> Activation_3.fcs|50%          1.5483953 1.2607698
#> Activation_3.fcs|75%          2.4766178 1.6177062
#> Activation_4.fcs|25%          1.0850130 0.8359168
#> Activation_4.fcs|50%          1.7589877 1.2615222
#> Activation_4.fcs|75%          2.5728453 1.6163138
#> Activation_5.fcs|25%          1.1055051 0.8117070
#> Activation_5.fcs|50%          1.7641329 1.2418383
#> Activation_5.fcs|75%          2.6454306 1.6393443
#> Activation_6.fcs|25%          1.0820792 0.8221035
#> Activation_6.fcs|50%          1.7452146 1.2747459
#> Activation_6.fcs|75%          2.6066504 1.6377774
#> Activation_7.fcs|25%          1.0501920 0.8523374
#> Activation_7.fcs|50%          1.7120249 1.3048321
#> Activation_7.fcs|75%          2.5958281 1.6907018
#> Activation_8.fcs|25%          1.0364495 0.8281519
#> Activation_8.fcs|50%          1.7051435 1.3047889
#> Activation_8.fcs|75%          2.6080720 1.7134358
#> Activation_9.fcs|25%          0.9893026 0.8561592
#> Activation_9.fcs|50%          1.5552643 1.2737321
#> Activation_9.fcs|75%          2.4305096 1.6300571
#> Activation_10.fcs|25%         0.9682094 0.8425388
#> Activation_10.fcs|50%         1.5899988 1.2800295
#> Activation_10.fcs|75%         2.4789634 1.6360306
#> Activation_11.fcs|25%         0.9863176 0.8202645
#> Activation_11.fcs|50%         1.5572237 1.2555786
#> Activation_11.fcs|75%         2.4218141 1.6294568
#> Activation_12.fcs|25%         1.0373567 0.8129501
#> Activation_12.fcs|50%         1.6217902 1.2674995
#> Activation_12.fcs|75%         2.4960406 1.6427241
#> Activation_13.fcs|25%         1.0043309 0.7953313
#> Activation_13.fcs|50%         1.5681664 1.2409685
#> Activation_13.fcs|75%         2.4594903 1.6228364
#> Activation_14.fcs|25%         1.0685717 0.7959368
#> Activation_14.fcs|50%         1.7165593 1.2554209
#> Activation_14.fcs|75%         2.5735731 1.6383598
#> Activation_15.fcs|25%         1.0405292 0.8418894
#> Activation_15.fcs|50%         1.6368322 1.3071570
#> Activation_15.fcs|75%         2.5256972 1.6969021
#> Activation_16.fcs|25%         1.0746655 0.8373898
#> Activation_16.fcs|50%         1.7106394 1.3321626
#> Activation_16.fcs|75%         2.6053114 1.7487255
#> Activation_17.fcs|25%         1.0723220 0.8632445
#> Activation_17.fcs|50%         1.6950802 1.2913077
#> Activation_17.fcs|75%         2.5443066 1.6539613
#> Activation_18.fcs|25%         1.0253232 0.8342279
#> Activation_18.fcs|50%         1.6827383 1.2565440
#> Activation_18.fcs|75%         2.5073639 1.5924325
#> Activation_19.fcs|25%         1.0250121 0.8228651
#> Activation_19.fcs|50%         1.6167989 1.2986770
#> Activation_19.fcs|75%         2.4848681 1.6780444
#> Activation_20.fcs|25%         1.0137688 0.8143133
#> Activation_20.fcs|50%         1.6198568 1.2604254
#> Activation_20.fcs|75%         2.4665596 1.6199343
#> Activation_21.fcs|25%         1.0668384 0.7954790
#> Activation_21.fcs|50%         1.6616037 1.2443692
#> Activation_21.fcs|75%         2.4707195 1.6493102
#> Activation_22.fcs|25%         1.0035264 0.8169925
#> Activation_22.fcs|50%         1.5949075 1.2681682
#> Activation_22.fcs|75%         2.4543028 1.6614087
#> Activation_23.fcs|25%         0.9987887 0.8820263
#> Activation_23.fcs|50%         1.5811659 1.3724525
#> Activation_23.fcs|75%         2.4073652 1.7919922
#> Activation_24.fcs|25%         1.0497267 0.8462257
#> Activation_24.fcs|50%         1.6472310 1.3586491
#> Activation_24.fcs|75%         2.5203644 1.7598584
#> Activation_25.fcs|25%         0.9576747 0.8445576
#> Activation_25.fcs|50%         1.5554076 1.2831917
#> Activation_25.fcs|75%         2.4115970 1.6198575
#> Activation_26.fcs|25%         0.9647052 0.8115377
#> Activation_26.fcs|50%         1.5320883 1.2410836
#> Activation_26.fcs|75%         2.3740296 1.6221624
#> Activation_27.fcs|25%         0.9571503 0.8378797
#> Activation_27.fcs|50%         1.4086682 1.2213336
#> Activation_27.fcs|75%         2.2806270 1.5705346
#> Activation_28.fcs|25%         0.9390236 0.8252140
#> Activation_28.fcs|50%         1.3896407 1.1983405
#> Activation_28.fcs|75%         2.2754512 1.5423217
#> Activation_29.fcs|25%         1.0208319 0.8231751
#> Activation_29.fcs|50%         1.5529205 1.3090844
#> Activation_29.fcs|75%         2.3990179 1.7128583
#> Activation_30.fcs|25%         0.9503936 0.8051299
#> Activation_30.fcs|50%         1.4706016 1.2525889
#> Activation_30.fcs|75%         2.3453180 1.6341102
#> Activation_31.fcs|25%         1.0118387 0.9067384
#> Activation_31.fcs|50%         1.6181533 1.4044728
#> Activation_31.fcs|75%         2.5088282 1.8291317
#> Activation_32.fcs|25%         0.9727298 0.8998926
#> Activation_32.fcs|50%         1.5321006 1.3555062
#> Activation_32.fcs|75%         2.4425468 1.8067185
#> Activation_33.fcs|25%                NA        NA
#> Activation_33.fcs|50%                NA        NA
#> Activation_33.fcs|75%                NA        NA
#>