Extract data from cytosets, GatingHierarchies or GatingSets

cyto_data_extract will extract populations specified by parent and return a list of cytosets. cyto_data_extract can also optionally perform additional tasks on the extracted cytosets, such subsetting channels, inversing transformations and returning raw data matrices.

cyto_data_extract(
  x,
  parent = "root",
  select = NULL,
  copy = TRUE,
  format = "cytoset",
  channels = NULL,
  markers = FALSE,
  split = FALSE,
  trans = NA,
  inverse = FALSE
)

Arguments

x
object of class cytoset, GatingHierarchy or GatingSet.

parent
name of the parent population to extract from GatingHierarchy or GatingSet objects.

select
named list containing experimental variables to be used to select samples using cyto_select.

copy
logical indicating whether a deep copy of the extracted data should be returned, set to TRUE by default to ensure that any changes to the data will not affect the supplied data.

format
can be either "cytoframe", "cytoset" or "matrix" to indicate in what format the data should be returned in the lists, set to "cytoset" by default.

channels
names of the markers or channels for which data should be extracted, set to all channels by default.

markers
logical indicating whether the column names of the raw data matrices should be replaced with the marker names where possible, set to FALSE by default.

split
logical indicating whether cytosets should be split into cytosets containing a single sample each, set to FALSE by default.

trans
object of class transformerList required to inverse transformations applied to the data when inverse = TRUE.

inverse
logical indicating whether transformations applied to the data should be inversed, set to FALSE by default.

Value

either a list of cytosets, list of cytoframe lists or a list of raw data matrix lists.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(system.file("extdata/Activation-GatingSet",
                package = "CytoExploreRData"))

# Extract cytosets & inverse transform
cyto_data_extract(gs[[1]],
parent = c("CD4 T Cells", "CD8 T Cells"),
inverse = TRUE)

#> $`CD4 T Cells`
#> A cytoset with 1 samples.
#> 
#>   column names:
#>     FSC-A, FSC-H, FSC-W, SSC-A, SSC-H, SSC-W, Alexa Fluor 488-A, PE-A, PE-Texas Red-A, 7-AAD-A, PE-Cy7-A, Alexa Fluor 405-A, Alexa Fluor 430-A, Qdot 605-A, Alexa Fluor 647-A, Alexa Fluor 700-A, APC-Cy7-A, Time
#> 
#> 
#> $`CD8 T Cells`
#> A cytoset with 1 samples.
#> 
#>   column names:
#>     FSC-A, FSC-H, FSC-W, SSC-A, SSC-H, SSC-W, Alexa Fluor 488-A, PE-A, PE-Texas Red-A, 7-AAD-A, PE-Cy7-A, Alexa Fluor 405-A, Alexa Fluor 430-A, Qdot 605-A, Alexa Fluor 647-A, Alexa Fluor 700-A, APC-Cy7-A, Time
#> 
#> 

# Extract raw data matrices
mt <- cyto_data_extract(gs,
parent = "root",
format = "matrix",
channels = c("CD4", "CD8"))
mt[["Activation_1.fcs"]]

#> NULL