Export cytometry data for use in flowJo or Cytobank

Simply a wrapper around CytoML gatingset_to_cytobank and gatingset_to _flowjo to export your CytoExploreR analyses in a format accepted by these platforms. Users will need to have docker running on their computers to export flowJo workspace files.

cyto_export(x, save_as = NULL, ...)

Arguments

x
object of class GatingSet to export.

save_as
the name of a flowJo workspace file with .wsp extension or a Cytobank xml file with .xml extension.

...
additional arguments passed to gatingset_to_cytobank or gatingset_to_flowjo.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

if (FALSE) {
library(CytoExploreRData)
library(CytoML)

# Activation GatingSet
gs <- GatingSet(Activation)

# Compensation
gs <- cyto_compensate(gs)

# Transformations
gs <- cyto_transform(gs)

# Gating
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Export to cytobank xml
cyto_export(gs, "cytobank.xml")

# Export to flowjo workspace
cyto_export(gs, "flowjo.wsp")
}