Retrieve details about experimental groups

Retrieve details about experimental groups

cyto_groups(x, group_by = "all", details = FALSE)

Arguments

x
object of class cytoset or GatingSet.

group_by
names of cyto_details variables to use for grouping For more control over the group order, specify the factor levels for each variable in a list (e.g. list(Treatment = c("Stim-A","Stim-C","Stim-B", "Stim-D"))).

details
logical indicating whether the split experimental details should be returned instead of the group names, set to FALSE by default.

Value

names of experimental groups or a list of experiment details per experimental group.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Treatment & OVAConc group names
cyto_groups(gs,
group_by = list(Treatment = c("Stim-A", "Stim-B", "Stim-C", "Stim-D"),
OVAConc = c(0, 5, 50, 500)))

#>  [1] "Stim-A 0"   "Stim-A 5"   "Stim-A 50"  "Stim-A 500" "Stim-B 0"  
#>  [6] "Stim-B 5"   "Stim-B 50"  "Stim-B 500" "Stim-C 0"   "Stim-C 5"  
#> [11] "Stim-C 50"  "Stim-C 500" "Stim-D 0"   "Stim-D 5"   "Stim-D 50" 
#> [16] "Stim-D 500" "NA 0"      

# Treatment & OVAConc group details
cyto_groups(gs,
group_by = list(Treatment = c("Stim-A", "Stim-B", "Stim-C", "Stim-D"),
OVAConc = c(0, 5, 50, 500)),
details = TRUE)

#> $`Stim-A 0`
#>                              name Treatment OVAConc
#> Activation_1.fcs Activation_1.fcs    Stim-A       0
#> Activation_2.fcs Activation_2.fcs    Stim-A       0
#> 
#> $`Stim-A 5`
#>                              name Treatment OVAConc
#> Activation_3.fcs Activation_3.fcs    Stim-A       5
#> Activation_4.fcs Activation_4.fcs    Stim-A       5
#> 
#> $`Stim-A 50`
#>                              name Treatment OVAConc
#> Activation_5.fcs Activation_5.fcs    Stim-A      50
#> Activation_6.fcs Activation_6.fcs    Stim-A      50
#> 
#> $`Stim-A 500`
#>                              name Treatment OVAConc
#> Activation_7.fcs Activation_7.fcs    Stim-A     500
#> Activation_8.fcs Activation_8.fcs    Stim-A     500
#> 
#> $`Stim-B 0`
#>                                name Treatment OVAConc
#> Activation_9.fcs   Activation_9.fcs    Stim-B       0
#> Activation_10.fcs Activation_10.fcs    Stim-B       0
#> 
#> $`Stim-B 5`
#>                                name Treatment OVAConc
#> Activation_11.fcs Activation_11.fcs    Stim-B       5
#> Activation_12.fcs Activation_12.fcs    Stim-B       5
#> 
#> $`Stim-B 50`
#>                                name Treatment OVAConc
#> Activation_13.fcs Activation_13.fcs    Stim-B      50
#> Activation_14.fcs Activation_14.fcs    Stim-B      50
#> 
#> $`Stim-B 500`
#>                                name Treatment OVAConc
#> Activation_15.fcs Activation_15.fcs    Stim-B     500
#> Activation_16.fcs Activation_16.fcs    Stim-B     500
#> 
#> $`Stim-C 0`
#>                                name Treatment OVAConc
#> Activation_17.fcs Activation_17.fcs    Stim-C       0
#> Activation_18.fcs Activation_18.fcs    Stim-C       0
#> 
#> $`Stim-C 5`
#>                                name Treatment OVAConc
#> Activation_19.fcs Activation_19.fcs    Stim-C       5
#> Activation_20.fcs Activation_20.fcs    Stim-C       5
#> 
#> $`Stim-C 50`
#>                                name Treatment OVAConc
#> Activation_21.fcs Activation_21.fcs    Stim-C      50
#> Activation_22.fcs Activation_22.fcs    Stim-C      50
#> 
#> $`Stim-C 500`
#>                                name Treatment OVAConc
#> Activation_23.fcs Activation_23.fcs    Stim-C     500
#> Activation_24.fcs Activation_24.fcs    Stim-C     500
#> 
#> $`Stim-D 0`
#>                                name Treatment OVAConc
#> Activation_25.fcs Activation_25.fcs    Stim-D       0
#> Activation_26.fcs Activation_26.fcs    Stim-D       0
#> 
#> $`Stim-D 5`
#>                                name Treatment OVAConc
#> Activation_27.fcs Activation_27.fcs    Stim-D       5
#> Activation_28.fcs Activation_28.fcs    Stim-D       5
#> 
#> $`Stim-D 50`
#>                                name Treatment OVAConc
#> Activation_29.fcs Activation_29.fcs    Stim-D      50
#> Activation_30.fcs Activation_30.fcs    Stim-D      50
#> 
#> $`Stim-D 500`
#>                                name Treatment OVAConc
#> Activation_31.fcs Activation_31.fcs    Stim-D     500
#> Activation_32.fcs Activation_32.fcs    Stim-D     500
#> 
#> $`NA 0`
#>                                name Treatment OVAConc
#> Activation_33.fcs Activation_33.fcs        NA       0
#>