Create dimension-reduced maps of cytometry data

Create dimension-reduced maps of cytometry data

cyto_map(x, ...)

# S3 method for GatingSet
cyto_map(
  x,
  parent = "root",
  select = NULL,
  channels = NULL,
  display = 1,
  type = "UMAP",
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)

# S3 method for flowSet
cyto_map(
  x,
  select = NULL,
  channels = NULL,
  display = 1,
  type = "UMAP",
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)

# S3 method for flowFrame
cyto_map(
  x,
  channels = NULL,
  display = 1,
  type = "UMAP",
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)

Arguments

x
object of class flowFrame or flowSet.

...
additional arguments passed to the called dimension reduction function. Links to the documentation for these functions can be found below.

parent
name of the parent population to extract from GatingHierarchy or GatingSet objects for mapping, set to the "root" node by default.

select
designates which samples should be used for mapping when a flowSet or GatingSet object is supplied. Filtering steps should be comma separated and wrapped in a list. Refer to cyto_select.

channels
vector of channels names indicating the channels that should be used by the dimension reduction algorithm to compute the 2-dimensional map, set to all channels with assigned markers by default. Restricting the number of channels can greatly improve processing speed and resolution.

display
total number of events to map, all events in the combined data are mapped by default.

type
dimension reduction type to use to generate the map, supported options include "PCA", "tSNE", "UMAP" and "EmbedSOM".

split
logical indicating whether samples merged using cyto_merge_by should be split prior to writing fcs files, set to FALSE by default.

names
original names of the samples prior to merging using cyto_merge_by, only required when split is TRUE. These names will be re-assigned to each of split flowFrames and included in the file names.

save_as
passed to cyto_save to indicate a folder where the mapped FCS files should be saved, set to NULL by default to turn off saving of FCS files.

inverse
logical indicating whether the data should be inverse transformed prior to writing FCS files, set to FALSE by default. Inverse transformations of flowFrame or flowSet objects requires passing of transformers through the trans argument.

trans
object of class transformerList containing the transformation definitions applied to the supplied data. Used internally when inverse_transform is TRUE, to inverse the transformations prior to writing FCS files.

plot
logical indicating whether the constructed map should be plotted using cyto_plot.

seed
integer to set seed prior to mapping to ensure more consistent results between runs.

Value

flowFrame, flowSet, GatingHierarchy or GatingSet containing the mapped projection parameters.

See also