Parse file names into experiment details

Reward users that use consistent names for files by parsing file names into new variables and adding them to cyto_details. Variable construction may not be perfect if certain files are named differently, but it will give a starting point for editing these values using cyto_details_edit without having to manually enter a lot of these details.

cyto_names_parse(x, vars = NULL, split = "_", exclude = NULL, ...)

Arguments

x
object of class cytoset or GatingSet.

vars
vector containing the names of the variables to be added to cyto_details, set to var_1, var_2 ... by default.

split
delimiter to split the file name into fragments, set to "_" by default.

exclude
vector of indices indicating which text fragments should not be included in the experiment details, set to NULL by default.

...
additional arguments passed to strsplit.

Value

cytoset or GatingSet with cyto_details updated with new variables extracted from the file names.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )
                          
# Parse file names to variables
gs <- cyto_names_parse(
  gs,
  vars = c("sample_type", "sample_id"),
  split = "_"
  )

# Updated experiment details
cyto_details(gs)

#>                   Treatment OVAConc              name sample_id sample_type
#> Activation_1.fcs     Stim-A       0  Activation_1.fcs         1  Activation
#> Activation_2.fcs     Stim-A       0  Activation_2.fcs         2  Activation
#> Activation_3.fcs     Stim-A       5  Activation_3.fcs         3  Activation
#> Activation_4.fcs     Stim-A       5  Activation_4.fcs         4  Activation
#> Activation_5.fcs     Stim-A      50  Activation_5.fcs         5  Activation
#> Activation_6.fcs     Stim-A      50  Activation_6.fcs         6  Activation
#> Activation_7.fcs     Stim-A     500  Activation_7.fcs         7  Activation
#> Activation_8.fcs     Stim-A     500  Activation_8.fcs         8  Activation
#> Activation_9.fcs     Stim-B       0  Activation_9.fcs         9  Activation
#> Activation_10.fcs    Stim-B       0 Activation_10.fcs        10  Activation
#> Activation_11.fcs    Stim-B       5 Activation_11.fcs        11  Activation
#> Activation_12.fcs    Stim-B       5 Activation_12.fcs        12  Activation
#> Activation_13.fcs    Stim-B      50 Activation_13.fcs        13  Activation
#> Activation_14.fcs    Stim-B      50 Activation_14.fcs        14  Activation
#> Activation_15.fcs    Stim-B     500 Activation_15.fcs        15  Activation
#> Activation_16.fcs    Stim-B     500 Activation_16.fcs        16  Activation
#> Activation_17.fcs    Stim-C       0 Activation_17.fcs        17  Activation
#> Activation_18.fcs    Stim-C       0 Activation_18.fcs        18  Activation
#> Activation_19.fcs    Stim-C       5 Activation_19.fcs        19  Activation
#> Activation_20.fcs    Stim-C       5 Activation_20.fcs        20  Activation
#> Activation_21.fcs    Stim-C      50 Activation_21.fcs        21  Activation
#> Activation_22.fcs    Stim-C      50 Activation_22.fcs        22  Activation
#> Activation_23.fcs    Stim-C     500 Activation_23.fcs        23  Activation
#> Activation_24.fcs    Stim-C     500 Activation_24.fcs        24  Activation
#> Activation_25.fcs    Stim-D       0 Activation_25.fcs        25  Activation
#> Activation_26.fcs    Stim-D       0 Activation_26.fcs        26  Activation
#> Activation_27.fcs    Stim-D       5 Activation_27.fcs        27  Activation
#> Activation_28.fcs    Stim-D       5 Activation_28.fcs        28  Activation
#> Activation_29.fcs    Stim-D      50 Activation_29.fcs        29  Activation
#> Activation_30.fcs    Stim-D      50 Activation_30.fcs        30  Activation
#> Activation_31.fcs    Stim-D     500 Activation_31.fcs        31  Activation
#> Activation_32.fcs    Stim-D     500 Activation_32.fcs        32  Activation
#> Activation_33.fcs        NA       0 Activation_33.fcs        33  Activation