Set customised graphical parameters for cyto_plot

cyto_plot_custom() make a call to cyto_plot_par() to set customised graphical parameters for the current graphics device. cyto_plot_custom() will then signal to cyto_plot() to ensure that these set graphical parameters are not overridden and to make sure that the plots are only saved when the user calls cyto_plot_complete(). cyto_plot_custom() will also record the current graphical parameters before changing them, so that these can be reset when cyto_plot_complete() is called.

cyto_plot_custom(...)

Arguments

...
graphical parameters supplied by name to be passed to cyto_plot_par to customise the current graphics device.

Details

The sequence of calls to save a customised plot are described below:

cyto_plot_custom() - sets up the current graphics device with customised graphical parameters and indicates to cyto_plot() that the user will take control over when the plot should be saved.

cyto_plot() - the users can then makes calls to any of the cyto_plot() family of function to construct the plot on the customised graphics device.

cyto_plot_complete() - signals that the plot is complete and ready for saving to file. Since cyto_plot_complete() also resets a lot of graphical parameters it should be called even if a call has not been made to cyto_plot_save().

Author

Dillon Hammill (Dillon.Hammill@anu.edu.au)

Examples

if (FALSE) {
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Save plot
cyto_plot_save("Test.png",
  height = 7,
  width = 14
)

# Create custom plot - 1D & 2D plot panels
cyto_plot_custom(layout = c(1, 2))
cyto_plot(fs[[32]],
  channels = "FSC-A"
)
cyto_plot(fs[[32]],
  channels = c("FSC-A", "SSC-A")
)

# Signal plot is complete and save
cyto_plot_complete()
}