Add density distributions to empty plots

Add density distributions to empty plots

cyto_plot_density(x, ...)

# S3 method for flowFrame
cyto_plot_density(
  x,
  channel,
  overlay = NA,
  display = 1,
  density_modal = TRUE,
  density_smooth = 1.5,
  density_stack = 0.5,
  density_cols = NA,
  density_fill = NA,
  density_fill_alpha = 1,
  density_line_type = 1,
  density_line_width = 1,
  density_line_col = "black",
  ...
)

# S3 method for list
cyto_plot_density(
  x,
  density_cols = NA,
  density_fill = NA,
  density_fill_alpha = 1,
  density_line_type = 1,
  density_line_width = 1,
  density_line_col = "black",
  ...
)

Arguments

x
list of density objects for plotting

...
not in use.

channel
name of the channels to be used to construct the plot.

overlay
list of flowFRame objects to overlay.

display
controls the number or percentage of events to display, set to 1 by default to display all events.

density_modal
logical indicating whether density should be normalised to mode and presented as a percentage. Set to TRUE by default.

density_smooth
smoothing parameter passed to density to adjust kernel density.

density_stack
numeric [0,1] indicating the degree of offset for overlaid populations, set to 0.5 by default.

density_cols
vector of colls to use to select density_fill colours

density_fill
colour(s) used to fill polygons.

density_fill_alpha
numeric [0,1] used to control fill transparency, set to 1 by default to remove transparency.

density_line_type
line type(s) to use for border(s), set to solid lines by default.

density_line_width
line width for border.

density_line_col
colour(s) for border line, set to "black" by default.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au