Create an empty cyto_plot

cyto_plot_empty generates to base for cyto_plot by creating an empty plot with border, axes, axes_text and titles. Data is subsequently added to this base layer with cyto_plot_point or cyto_plot_hist.

cyto_plot_empty(
  x,
  channels,
  axes_trans = NA,
  overlay = NA,
  gate = NA,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes_limits = "auto",
  axes_limits_buffer = 0.03,
  title,
  xlab,
  ylab,
  margins = c(NA, NA, NA, NA),
  hist_stat = "count",
  hist_smooth = 1,
  hist_bins = 256,
  hist_stack = 0.5,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  point_shape = ".",
  point_size = 2,
  point_col_scale = NA,
  point_cols = NA,
  point_col = NA,
  point_col_alpha = 1,
  axes_text = c(TRUE, TRUE),
  axes_text_font = 1,
  axes_text_size = 1,
  axes_text_col = "black",
  axes_label_text_font = 1,
  axes_label_text_size = 1.1,
  axes_label_text_col = "black",
  title_text_font = 2,
  title_text_size = 1.1,
  title_text_col = "black",
  border_line_type = 1,
  border_line_width = 1,
  border_line_col = "black",
  border_fill = "white",
  border_fill_alpha = 1,
  legend = FALSE,
  legend_text,
  legend_text_font = 1,
  legend_text_size = 1,
  legend_text_col = "black",
  legend_line_type = NA,
  legend_line_width = NA,
  legend_line_col = NA,
  legend_box_fill = NA,
  legend_point_col = NA,
  grid = TRUE,
  grid_line_type = 1,
  grid_line_width = 1,
  grid_line_col = "grey95",
  grid_line_alpha = 1,
  ...
)

Arguments

x
object of class flowFrame.

channels
name of the channel(s) or marker(s) to be used to construct the plot. The length of channels determines the type of plot to be constructed, either a 1-D density distribution for a single channel or a 2-D scatterplot with blue-red colour scale for two channels.

axes_trans
object of class transformList or transformerList generated by estimateLogicle which was used to transform the fluorescent channels of the supplied flowFrame. This transformation object will be used internally to ensure that the axes labels of the plot are appropriately transformed. The transformation object will NOT be applied to the flowFrame internally and should be applied to the flowFrame prior to plotting.

overlay
a list of flowFrames to overlay onto the plot.

gate
list of gate objects to be plotted, used internlaly to ensure gate co-ordinates are taken into account when computing axes limits.

xlim
lower and upper limits of x axis (e.g. c(0,5)).

ylim
lower and upper limits of y axis (e.g. c(0,5)).

axes_limits
options include "auto", "data" or "machine" to use optimised, data or machine limits respectively. Set to "auto" by default to use optimised axes ranges. Fine control over axes limits can be obtained by altering the xlim and ylim arguments.

axes_limits_buffer
decimal indicating the percentage of buffering to add to either end of the axes limits, set to 0.03 by default.

title
title to use for the plot, set to the name of the sample by default. Title can be removed by setting this argument to NA.

xlab
x axis label.

ylab
y axis label.

margins
a vector of length 4 to control the margins around the bottom, left, top and right of the plot, set to c(NA, NA, NA, NA) by default to let `cyto_plot` compute optimal margins.

hist_stat
can be either "count", "percent" or "density" to indicate the statistic to display on histograms, set to "percent" by default. The "percent" option applies modal normalisation and expresses the result as a percentage.

hist_smooth
smoothing parameter passed to density to adjust the smoothness of the kernel density for histograms, set to 1 by default.

hist_bins
number of bins to use for histograms, set to 256 by default.

hist_stack
numeric [0,1] indicating the degree of stacking for histograms, set to 0 by default.

hist_cols
vector colours to draw from when selecting histogram fill colours if none are supplied to hist_fill.

hist_fill
fill colour(s) for histograms, select from hist_cols if not supplied.

hist_fill_alpha
numeric [0,1] used to control histogram fill colour transparency, set to 1 by default for solid colours.

hist_line_type
line type(s) to use for histogram borders, set to 1 by default to use solid lines. See lty for alternatives.

hist_line_width
numeric to control line width(s) for histogram borders lines, set to 1 by default.

hist_line_col
colour(s) for histogram borders, set to "black" by default.

point_shape
shape(s) to use for points in 2-D scatterplots, set to "." by default to maximise plotting speed. See pch for alternatives.

point_size
numeric to control the size of points in 2-D scatter plots set to 2 by default.

point_col_scale
vector of colours to use for density gradient.

point_cols
vector colours to draw from when selecting colours for points if none are supplied to point_col.

point_col
colour(s) to use for points in 2-D scatter plots, set to NA by default to use a blue-red density colour scale.

point_col_alpha
numeric [0,1] to control point colour transparency in 2-D scatter plots, set to 1 by default to use solid colours.

axes_text
logical vector of length 2 indicating whether axis text should be included for the x and y axes respectively, set to c(TRUE,TRUE) by default to display axes text on both axes.

axes_text_font
numeric indicating the font to use for axes, set to 1 for plain font by default. See ?par font for details.

axes_text_size
character expansion for axis text.

axes_text_col
colour of axis text.

axes_label_text_font
numeric indicating the font to use for title, set to 1 for plain font by default. See ?par font for details.

axes_label_text_size
character expansion for axis labels.

axes_label_text_col
colour of axis labels.

title_text_font
numeric indicating the font to use for title, set to 2 for bold font by default. See ?par font for details.

title_text_size
character expansion for plot title.

title_text_col
colour for plot title.

border_line_type
line type to use for plot border, set to 1 by default for a sold border.

border_line_width
line width for plot border, set to 1 by default.

border_line_col
line colour for plot border, set to "black" by default.

border_fill
colour to use for the plot background, set to "white" by default.

border_fill_alpha
transparency to use for border_fill colour, set to 1 by default to add no transparency.

legend
can be either "line" or "fill" to indicate whether a legend should be constructed based on the density "line" or "fill", set to FALSE by default to remove the legend.

legend_text
vector of labels to use in the legend.

legend_text_font
numeric to control the font of legend text, set to 1 for plain font by default. See font for alternatives.

legend_text_size
numeric to control the size of text in the legend, set to 1 by default.

legend_text_col
colour(s) to use for text in legend, set to "black" by default.

legend_line_type
numeric to control the line type for line legends, set to 1 by default. Refer to lty in par for alternatives.

legend_line_width
numeric to control the line width in line legend, set to 1 by default. Refer to lwd in par for alternatives.

legend_line_col
colour(s) to use for the lines in 1-D plot legends when legend is set to "line".

legend_box_fill
fill colour(s) to use for the boxes in 1-D plot legends when legend is set to "fill".

legend_point_col
colour(s) to use for points in 2-D scatter plot legend.

grid
logical indicating whether to include grid lines in the plot background, set to TRUE by default. Alternatively, users can supply a integer to indicate the number of equally spaced quantiles to used for the grid lines.

grid_line_type
integer [0,6] to control the line type of grid lines, set to 1 to draw solid lines by default. See lty for alternatives.

grid_line_width
numeric to control the line width(s) of grid lines, set to 1 by default.

grid_line_col
colour to use for grid lines, set to "grey95" by default.

grid_line_alpha
numeric [0,1] to control the transparency of grid lines, set to 1 by default to remove transparency.

...
not in use.

hist_layers
numeric indicating the number of histograms to stack in each plot, set to all samples by default. Each plot must contain the same number of histograms.

Author

Dillon Hammill (Dillon.Hammill@anu.edu.au)

Examples

library(CytoExploreRData)

# Construct an empty 2D plot with black background
cyto_plot_empty(Activation[[32]],
  channels = c("FSC-A", "SSC-A"),
  border_fill = "black"
)

#> Error in UseMethod("cyto_names"): no applicable method for 'cyto_names' applied to an object of class "logical"

# Construct an empty 1D plot
cyto_plot_empty(Activation[[32]],
  channels = c("FSC-A"),
  overlay = Activation[1:2]
)

#> Error in cyto_apply.default(cs, "cyto_stat_density", input = "matrix",     channels = channels, copy = FALSE, stat = hist_stat, bins = hist_bins,     smooth = hist_smooth, limits = matrix(xlim, ncol = 1, dimnames = list(c("min",         "max"), channels)), simplify = FALSE): 'cyto_apply' does not accept objects of class flowFrame!