Explore Cytometry Data in Bivariate Plots

cyto_plot_explore is an extremely useful tool to explore all aspectes of your cytometry data. cyto_plot_explore constructs a faceted plot for each of the supplied channels_x by plotting them against all the supplied channels_y. Thus providing a rapid means tro explore the data in all available channels.

# S3 method for GatingSet
cyto_plot_explore(
  x,
  parent = NULL,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

# S3 method for GatingHierarchy
cyto_plot_explore(
  x,
  parent = NULL,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

# S3 method for flowSet
cyto_plot_explore(
  x,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

# S3 method for flowFrame
cyto_plot_explore(
  x,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

Arguments

x
an object of class flowFrame, flowSet, GatingHierachy or GatingSet. flowSet and GatingSet objects will be coerced to a single flowFrame prior to plotting.

parent
name of the parent population to plot when a GatingHierarchy or GatingSet object is supplied, set to the last gated population by default.

channels_x
vector of channels to explore, set to all fluorescent channels by default. Each channel in channels_x will be plotted against all channels_y, with each channels_x on a separate page.

channels_y
vector of channels to plot each channels_x against, set to all channels by default.

axes_trans
object of class transformerList passed to cyto_plot to appropriately display transformed data in plots.

layout
vector of grid dimensions c(#rows,#columns) for each plot.

popup
logical indicating whether plots should be constructed in a pop-up window.

title
text to include above each plot, set to NA by default to remove titles.

header
title to use for the plots, set to the name of the sample by default. Turn off the header by setting this argument to NA.

header_text_font
font to use for header text, set to 2 by default.

header_text_size
text size for header, set to 1 by default.

header_text_col
colour for header text, set to "black" by default.

...
additional arguments passed to cyto_plot.