Add histograms to empty plots

Add histograms to empty plots

cyto_plot_hist(
  x,
  parent = "root",
  channel,
  overlay = NA,
  display = 1,
  hist_stat = "count",
  hist_bins = 256,
  hist_smooth = 1,
  hist_stack = 0.5,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  seed = 42,
  ...
)

Arguments

x
object of class cytoframe, cytoset, list of cytoframes or a list of density objects.

parent
name of the parent population to extract from GatingHierarchy or GatingSet objects.

channel
name of the channels to be used to construct the plot.

overlay
list of cytoframe objects to overlay.

display
controls the number or percentage of events to display, set to 1 by default to display all events.

hist_stat
can be either "count", "percent" or "density" to indicate the statistic to display on histograms, set to "count" by default. The "percent" option applies modal normalisation and expresses the result as a percentage.

hist_bins
number of bins to use for histograms, set to 256 by default.

hist_smooth
smoothing parameter passed to density to adjust the smoothness of the kernel density for histograms, set to 1 by default. Only values greater or equal to 1 are supported.

hist_stack
numeric [0,1] indicating the degree of stacking for histograms, set to 0.5 by default.

hist_cols
vector colours to draw from when selecting histogram fill colours if none are supplied to hist_fill.

hist_fill
fill colour(s) for histograms, select from hist_cols if not supplied.

hist_fill_alpha
numeric [0,1] used to control histogram fill colour transparency, set to 1 by default for solid colours.

hist_line_type
line type(s) to use for histogram borders, set to 1 by default to use solid lines. See lty for alternatives.

hist_line_width
numeric to control line width(s) for histogram borders lines, set to 1 by default.

hist_line_col
colour(s) for histogram borders, set to "black" by default.

seed
numeric passed to set.seed to ensure that the same sampling is applied with each cyto_plot_contour call, set to an arbitrary numeric by default. This behaviour can be turned off by setting this argument to NULL.

...
not in use.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au