Plot expression profile in all channels

Plot expression profile in all channels

# S3 method for GatingSet
cyto_plot_profile(
  x,
  parent = NULL,
  channels = NULL,
  group_by = NA,
  select = NULL,
  axes_trans = NA,
  layout = NULL,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  ...
)

# S3 method for GatingHierarchy
cyto_plot_profile(
  x,
  parent = NULL,
  channels = NULL,
  axes_trans = NA,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  ...
)

# S3 method for flowSet
cyto_plot_profile(
  x,
  channels = NULL,
  group_by = NA,
  select = NA,
  axes_trans = NULL,
  layout = NULL,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  density_stack = 0.5,
  ...
)

# S3 method for flowFrame
cyto_plot_profile(
  x,
  channels = NULL,
  axes_trans = NA,
  layout = NULL,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  ...
)

Arguments

x
object of class flowFrame, flowSet, GatingHierarchy, GatingSet.

parent
name of the population to plot when a GatingHierarchy or GatingSet object is supplied.

channels
a vector channels to use to construct the plots, set to all channels by default.

group_by
a vector of experiment variables to sort and merge samples into groups prior to plotting, set to NA by default to prevent merging. To merge all samples set this argument to "all".

select
designates which samples will be plotted and used for determining the best location to set the drawn gate(s). Filtering steps should be comma separated and wrapped in a list. Refer to cyto_select.

axes_trans
object of class transformerList which was used to transform the channels of the supplied flowFrame. cyto_plot does not support in-line transformations and as such the transformations should be applied to the data prior to plotting. The transformerList is used internally to ensure that the axes on the constructed plots are appropriately labelled.

layout
a vector of the length 2 indicating the dimensions of the grid for plotting c(#rows, #columns).

header
character string to include in the plot header, set to the sample names by deafult. The header can be removed by setting this argument to NA.

header_text_font
numeric indicating the font to use for the header, set to 2 by default for bold font.

header_text_size
numeric to control the text size of the header, set to 1 by deafult.

...
additional arguments passed to cyto_plot.

density_stack
numeric [0,1] indicating the degree of offset for 1-D density distributions with overlay, set to 0.5 by default.

See also

Examples

# Load in CytoExploreR to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
gs <- cyto_transform(gs, trans_type = "logicle")

#> Insufficient plotting space, transformations have been applied.

# Apply gates
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

#> Preprocessing for 'Cells'

#> Gating for 'Cells'

#> done!

#> done.

#> Preprocessing for 'Single Cells'

#> Gating for 'Single Cells'

#> done!

#> done.

#> Preprocessing for 'Dead Cells'

#> Gating for 'Dead Cells'

#> done!

#> done.

#> Live Cells gating...

#> done!

#> done.

#> Preprocessing for 'T Cells'

#> Gating for 'T Cells'

#> done!

#> done.

#> Preprocessing for 'CD8 T Cells'

#> Gating for 'CD8 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD69+ CD8 T Cells'

#> Gating for 'CD69+ CD8 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD4 T Cells'

#> Gating for 'CD4 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD69+ CD4 T Cells'

#> Gating for 'CD69+ CD4 T Cells'

#> done!

#> done.

#> Preprocessing for 'Dendritic Cells'

#> Gating for 'Dendritic Cells'

#> done!

#> done.

#> finished.

# Plot expression profile of T Cells in all channels
cyto_plot_profile(gs[1:9],
  parent = "T Cells"
)

#> Error in plot.new(): figure margins too large

# Group samples by Treatment & select highest OVAConc
 cyto_plot_profile(gs[1:9],
  parent = "CD4 T Cells",
  group_by = "Treatment",
  select = list("OVAConc" = 500),
)

#> Error in plot.new(): figure margins too large