Compute, export and save statistics

Compute, export and save statistics

cyto_stats_compute(
  x,
  alias = NULL,
  parent = NULL,
  channels = NULL,
  trans = NA,
  stat = NULL,
  inverse = TRUE,
  round = 2,
  format = "wide",
  tibble = FALSE,
  input = "matrix",
  smooth = 1,
  bins = 256,
  details = TRUE,
  markers = TRUE,
  save_as = NULL,
  select = NULL,
  ...
)

Arguments

x
object of class cytoset, GatingHierarchy or GatingSet.

alias
names of the populations in the GatingHerarchy or GatingSet for which statistics should be computed.

parent
names of the parent populations in the GatingHierarchy or GatingSet required when computing frequency statistics. When stat = "freq" the frequency of each alias population will be computed as a proportion of each parent population.

channels
names of the channels or markers for which statistics should be computed, set to all channels with marker assignments by default.

trans
an object of class transformerList containing the transformers used to transform the channels of the supplied data. The transformerList will be automatically extracted from the GatingHierarchy or GatingSet, so this argument is only required for cytoframes or cytosets. The transformerList is passed to cyto_apply which will make sure the data is appropriately transformed based on inverse prior to passing the data to the desired statistical function.

stat
name of a statistic to compute, options include:

count - number of events

freq/percent - frequency of population(s) in parent population(s)

mean - arithmetic mean

geomean - geometric (graphical) mean computed as the inverse of the arithmetic mean on the transformed scale

median - 50th quantile

mode - inverse of mode on transformed scale computed using smooth and bandwidth to control the smoothness of the density distribution

sd - standard deviation

rsd - robust standard deviation (less influence of outliers)

cv - coefficient of variation

rcv - robust coefficient of variation (less influence of outliers)

quantile - quantiles with probabilities supplied to probs

range - minima and maxima

auc - area under curve using a combination of density(), splinefun() and integrate()

stat dispatches through cyto_apply so any custom function can be named through this argument as well. See cyto_apply for details.

inverse
logical passed to cyto_apply to indicate whether transformations applied to the data should be reversed prior to passing the data to the desired statistical function, set to TRUE by default.

round
numeric indicating the number of decimal places to round the computed statistic, set to 2 decimal places by default.

format
can be either "wide" or "long" to control the format of the returned statistics, set to "wide" by default.

tibble
logical indicating whether the statistics should be returned as tibble instead of a data.frame, set to FALSE by default.

input
passed to cyto_apply to control how the data is formatted prior to passing it to the statistical function, options include:

1 - "cytoframe"

2 - "matrix"

3 - "column" or "channel"

4 - "row" or "cell"

Set to "matrix" by default.

smooth
numeric smoothing parameter passed to stats:density when computing mode and area under the curve statistics, set to 1 by default.

bins
number of bins to use for histograms, set to 256 by default.

details
logical indicating whether to include the cyto_details in the output, set to TRUE by default.

markers
logical indicating whether channels should be converted to markers where possible in the output, set to TRUE by default.

save_as
name of a csv file to which the output should be saved, set to NULL by default to bypass this saving step.

select
named list containing experimental variables to be used to select samples using cyto_select when a cytoset or GatingSet is supplied. Refer to cyto_select for more details.

...
additional arguments passed to the desired statistical function.

gate
a rectangleGate, polygonGate or ellipsoidGate object to apply to each cytoframe prior to computing the desired statistic. This argument has been included for backwards compatibility only and is only valid for cytoframe or cytoset objects.

Value

data.frame or tibble containing the computed statistics in the desired format and optionally a csv file containing the computed statistics.

Author

Dillon Hammill, Dillon.Hammill@anu.edu.au

Examples

library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Compensation
gs <- cyto_compensate(gs)

# Transformations
gs <- cyto_transform(gs)

#> Insufficient plotting space, transformations have been applied.

# Gating
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

#> Preprocessing for 'Cells'

#> Gating for 'Cells'

#> done!

#> done.

#> Preprocessing for 'Single Cells'

#> Gating for 'Single Cells'

#> done!

#> done.

#> Preprocessing for 'Dead Cells'

#> Gating for 'Dead Cells'

#> done!

#> done.

#> Live Cells gating...

#> done!

#> done.

#> Preprocessing for 'T Cells'

#> Gating for 'T Cells'

#> done!

#> done.

#> Preprocessing for 'CD8 T Cells'

#> Gating for 'CD8 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD69+ CD8 T Cells'

#> Gating for 'CD69+ CD8 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD4 T Cells'

#> Gating for 'CD4 T Cells'

#> done!

#> done.

#> Preprocessing for 'CD69+ CD4 T Cells'

#> Gating for 'CD69+ CD4 T Cells'

#> done!

#> done.

#> Preprocessing for 'Dendritic Cells'

#> Gating for 'Dendritic Cells'

#> done!

#> done.

#> finished.

# Compute statistics - median
cyto_stats_compute(gs,
  alias = "T Cells",
  channels = c("Alexa Fluor 488-A", "PE-A"),
  stat = "median"
)

#>                 name Treatment OVAConc   alias     CD8      Va2
#> 1   Activation_1.fcs    Stim-A       0 T Cells 1500.80 17650.05
#> 2   Activation_2.fcs    Stim-A       0 T Cells 2256.58 18609.77
#> 3   Activation_3.fcs    Stim-A       5 T Cells 1540.68 18946.23
#> 4   Activation_4.fcs    Stim-A       5 T Cells  567.54 19874.78
#> 5   Activation_5.fcs    Stim-A      50 T Cells 1102.45 20104.33
#> 6   Activation_6.fcs    Stim-A      50 T Cells 2164.58 20074.90
#> 7   Activation_7.fcs    Stim-A     500 T Cells 4808.94 20486.06
#> 8   Activation_8.fcs    Stim-A     500 T Cells 4370.65 20925.35
#> 9   Activation_9.fcs    Stim-B       0 T Cells 3354.17 17586.40
#> 10 Activation_10.fcs    Stim-B       0 T Cells 1297.17 17973.38
#> 11 Activation_11.fcs    Stim-B       5 T Cells 3857.94 18340.17
#> 12 Activation_12.fcs    Stim-B       5 T Cells 4350.15 18871.20
#> 13 Activation_13.fcs    Stim-B      50 T Cells 3621.18 18426.13
#> 14 Activation_14.fcs    Stim-B      50 T Cells 2291.32 19365.62
#> 15 Activation_15.fcs    Stim-B     500 T Cells 4850.92 19410.60
#> 16 Activation_16.fcs    Stim-B     500 T Cells 5083.20 20546.19
#> 17 Activation_17.fcs    Stim-C       0 T Cells  759.02 19795.24
#> 18 Activation_18.fcs    Stim-C       0 T Cells  478.81 18129.59
#> 19 Activation_19.fcs    Stim-C       5 T Cells 2509.31 19803.92
#> 20 Activation_20.fcs    Stim-C       5 T Cells  530.12 19371.80
#> 21 Activation_21.fcs    Stim-C      50 T Cells 4769.55 20884.91
#> 22 Activation_22.fcs    Stim-C      50 T Cells 4751.83 20005.57
#> 23 Activation_23.fcs    Stim-C     500 T Cells 5904.28 17984.60
#> 24 Activation_24.fcs    Stim-C     500 T Cells 6559.26 20215.38
#> 25 Activation_25.fcs    Stim-D       0 T Cells 4547.27 18816.81
#> 26 Activation_26.fcs    Stim-D       0 T Cells 4544.19 17962.56
#> 27 Activation_27.fcs    Stim-D       5 T Cells 4689.65 16231.62
#> 28 Activation_28.fcs    Stim-D       5 T Cells 4493.16 15694.07
#> 29 Activation_29.fcs    Stim-D      50 T Cells 6392.02 19035.28
#> 30 Activation_30.fcs    Stim-D      50 T Cells 5344.23 16617.94
#> 31 Activation_31.fcs    Stim-D     500 T Cells 6405.79 18495.34
#> 32 Activation_32.fcs    Stim-D     500 T Cells 6389.46 17582.23
#> 33 Activation_33.fcs        NA       0 T Cells     NaN      NaN

# Name csv to save results
tempfile <- paste0(tempdir(),
.Platform$file.sep,
"Population-Frequencies.csv")

# Compute frequencies and save to tempfile
cyto_stats_compute(gs,
  alias = c("CD4 T Cells", "CD8 T Cells"),
  parent = c("Live Cells", "T Cells"),
  stat = "freq",
  save_as = tempfile
)

#>                 name Treatment OVAConc       alias Live Cells T Cells
#> 1   Activation_1.fcs    Stim-A       0 CD4 T Cells      13.27   41.17
#> 2   Activation_2.fcs    Stim-A       0 CD4 T Cells      14.99   41.18
#> 3   Activation_3.fcs    Stim-A       5 CD4 T Cells      14.86   40.77
#> 4   Activation_4.fcs    Stim-A       5 CD4 T Cells      13.57   42.17
#> 5   Activation_5.fcs    Stim-A      50 CD4 T Cells      15.09   40.26
#> 6   Activation_6.fcs    Stim-A      50 CD4 T Cells      13.47   40.61
#> 7   Activation_7.fcs    Stim-A     500 CD4 T Cells      14.02   38.94
#> 8   Activation_8.fcs    Stim-A     500 CD4 T Cells      13.57   39.03
#> 9   Activation_9.fcs    Stim-B       0 CD4 T Cells      13.00   38.74
#> 10 Activation_10.fcs    Stim-B       0 CD4 T Cells      12.30   40.53
#> 11 Activation_11.fcs    Stim-B       5 CD4 T Cells      13.87   38.14
#> 12 Activation_12.fcs    Stim-B       5 CD4 T Cells      11.33   37.05
#> 13 Activation_13.fcs    Stim-B      50 CD4 T Cells      13.50   38.31
#> 14 Activation_14.fcs    Stim-B      50 CD4 T Cells      12.83   39.62
#> 15 Activation_15.fcs    Stim-B     500 CD4 T Cells      14.32   36.75
#> 16 Activation_16.fcs    Stim-B     500 CD4 T Cells      13.68   38.18
#> 17 Activation_17.fcs    Stim-C       0 CD4 T Cells      12.64   40.88
#> 18 Activation_18.fcs    Stim-C       0 CD4 T Cells      12.66   42.42
#> 19 Activation_19.fcs    Stim-C       5 CD4 T Cells      15.70   40.49
#> 20 Activation_20.fcs    Stim-C       5 CD4 T Cells      13.44   43.06
#> 21 Activation_21.fcs    Stim-C      50 CD4 T Cells      15.48   38.13
#> 22 Activation_22.fcs    Stim-C      50 CD4 T Cells      13.41   37.80
#> 23 Activation_23.fcs    Stim-C     500 CD4 T Cells      12.90   33.87
#> 24 Activation_24.fcs    Stim-C     500 CD4 T Cells      12.95   33.96
#> 25 Activation_25.fcs    Stim-D       0 CD4 T Cells      13.94   37.34
#> 26 Activation_26.fcs    Stim-D       0 CD4 T Cells      13.72   38.01
#> 27 Activation_27.fcs    Stim-D       5 CD4 T Cells      13.72   36.72
#> 28 Activation_28.fcs    Stim-D       5 CD4 T Cells      13.08   36.35
#> 29 Activation_29.fcs    Stim-D      50 CD4 T Cells      12.68   32.30
#> 30 Activation_30.fcs    Stim-D      50 CD4 T Cells      14.22   34.86
#> 31 Activation_31.fcs    Stim-D     500 CD4 T Cells      11.84   31.63
#> 32 Activation_32.fcs    Stim-D     500 CD4 T Cells      11.42   31.40
#> 33 Activation_33.fcs        NA       0 CD4 T Cells       0.00     NaN
#> 34  Activation_1.fcs    Stim-A       0 CD8 T Cells      16.05   49.78
#> 35  Activation_2.fcs    Stim-A       0 CD8 T Cells      18.30   50.29
#> 36  Activation_3.fcs    Stim-A       5 CD8 T Cells      18.12   49.72
#> 37  Activation_4.fcs    Stim-A       5 CD8 T Cells      15.39   47.83
#> 38  Activation_5.fcs    Stim-A      50 CD8 T Cells      18.45   49.23
#> 39  Activation_6.fcs    Stim-A      50 CD8 T Cells      16.64   50.20
#> 40  Activation_7.fcs    Stim-A     500 CD8 T Cells      19.01   52.78
#> 41  Activation_8.fcs    Stim-A     500 CD8 T Cells      18.00   51.79
#> 42  Activation_9.fcs    Stim-B       0 CD8 T Cells      17.35   51.72
#> 43 Activation_10.fcs    Stim-B       0 CD8 T Cells      15.00   49.44
#> 44 Activation_11.fcs    Stim-B       5 CD8 T Cells      19.19   52.73
#> 45 Activation_12.fcs    Stim-B       5 CD8 T Cells      16.26   53.19
#> 46 Activation_13.fcs    Stim-B      50 CD8 T Cells      18.41   52.24
#> 47 Activation_14.fcs    Stim-B      50 CD8 T Cells      16.44   50.74
#> 48 Activation_15.fcs    Stim-B     500 CD8 T Cells      20.97   53.79
#> 49 Activation_16.fcs    Stim-B     500 CD8 T Cells      19.14   53.41
#> 50 Activation_17.fcs    Stim-C       0 CD8 T Cells      15.02   48.56
#> 51 Activation_18.fcs    Stim-C       0 CD8 T Cells      14.12   47.31
#> 52 Activation_19.fcs    Stim-C       5 CD8 T Cells      19.61   50.58
#> 53 Activation_20.fcs    Stim-C       5 CD8 T Cells      14.88   47.69
#> 54 Activation_21.fcs    Stim-C      50 CD8 T Cells      21.52   53.02
#> 55 Activation_22.fcs    Stim-C      50 CD8 T Cells      18.93   53.35
#> 56 Activation_23.fcs    Stim-C     500 CD8 T Cells      21.39   56.17
#> 57 Activation_24.fcs    Stim-C     500 CD8 T Cells      21.72   56.96
#> 58 Activation_25.fcs    Stim-D       0 CD8 T Cells      19.94   53.42
#> 59 Activation_26.fcs    Stim-D       0 CD8 T Cells      19.33   53.57
#> 60 Activation_27.fcs    Stim-D       5 CD8 T Cells      20.46   54.76
#> 61 Activation_28.fcs    Stim-D       5 CD8 T Cells      19.99   55.56
#> 62 Activation_29.fcs    Stim-D      50 CD8 T Cells      22.96   58.47
#> 63 Activation_30.fcs    Stim-D      50 CD8 T Cells      23.05   56.48
#> 64 Activation_31.fcs    Stim-D     500 CD8 T Cells      21.61   57.73
#> 65 Activation_32.fcs    Stim-D     500 CD8 T Cells      21.51   59.17
#> 66 Activation_33.fcs        NA       0 CD8 T Cells       0.00     NaN