Definition(s) of biexponential transformation(s)

CytoExploreR implementation of flowWorkspace flowJo biexponential transformation which always returns a transformerList object and displays in the result of the transformation(s) using cyto_plot. To combine different types of transformations have a look at cyto_transformer_combine. cyto_transform should be used to apply the transformations to the data.

# S3 method for GatingSet
cyto_transformer_biex(
  x,
  channels = NULL,
  parent = "root",
  select = NULL,
  plot = TRUE,
  popup = FALSE,
  display = 25000,
  axes_limits = "machine",
  ...
)

# S3 method for GatingHierarchy
cyto_transformer_biex(
  x,
  channels = NULL,
  parent = "root",
  select = NULL,
  plot = TRUE,
  popup = FALSE,
  display = 25000,
  axes_limits = "machine",
  ...
)

# S3 method for flowSet
cyto_transformer_biex(
  x,
  channels = NULL,
  parent = "root",
  select = NULL,
  plot = TRUE,
  popup = FALSE,
  display = 25000,
  axes_limits = "machine",
  ...
)

# S3 method for flowFrame
cyto_transformer_biex(
  x,
  channels = NULL,
  parent = "root",
  select = NULL,
  plot = TRUE,
  popup = FALSE,
  display = 25000,
  axes_limits = "machine",
  ...
)

Arguments

x
an object of class flowFrame, flowSet, GatingHierarchy or GatingSet.

channels
name(s) of channel(s)/marker(s) for which transformation functions must be generated. Set to all the fluorescent channels by default if no channels/markers are supplied.

parent
name of the parent population to use for generating the transformation(s).

select
list of selection criteria passed to cyto_select to select a subset of samples for cyto_plot.

plot
logical indicating whether the results of transformations should be plotted using cyto_plot.

popup
logical indicating whether the plots should be constructed in a pop-up window.

display
number or frequency of events to display in plots, set to 25000 events by default. See cyto_plot for details.

axes_limits
options include "auto", "data" or "machine" to use optimised, data or machine limits respectively. Set to "machine" by default to use entire axes ranges. Fine control over axes limits can be obtained by altering the xlim and ylim arguments.

...
additional arguments passed to flowjo_biexp and cyto_plot.

Value

a transformerList object.

See also