Load.fcs files into GatingSet and annotate with experiment details

cyto_setup takes care of all the data loading and annotation steps to prepare your cytometry data for downstream analyses. The .fcs files are first read into a cytoset using cyto_load which is then added to a GatingSet.

cyto_setup(
  path = ".",
  gatingTemplate = NULL,
  restrict = FALSE,
  clean = FALSE,
  markers = TRUE,
  parse_names = FALSE,
  details = TRUE,
  ...
)

Arguments

path
points to the location of the .fcs files to read in (e.g. name of a folder in current working directory).

gatingTemplate
name of a gatingTemplate csv file to be used for gate saving.

restrict
logical indicating whether unassigned channels should be dropped from the returned cytoset, set to FALSE by default. Alternatvely, users can supply a vector of channels to remove, see cyto_channels_restrict for details.

clean
logical indicating whether the loaded data should be cleaned using cyto_clean, set to FALSE by default. Alternatively, users can indicate which types of anomalies should be checked as expected by remove_from in flow_auto_qc.

markers
logical indicating whether a call should be made to cyto_markers_edit to update the markers associated with channels in the loaded sampes, set to TRUE by default. The name of the csv to which these details will be supplied can also be passed to this argument.

parse_names
logical indicating whether the file names should be parsed into experiment details using cyto_names_parse, set to FALSE by default. If you need to parse the names using a different dlimiter, supply the delimiter to this argument instead of TRUE.

details
logical indicating whether a call should be made to cyto_details_edit to update the experimental details associated with the loaded samples, set to TRUE by default. The name of the csv to which these details will be supplied can also be passed to this argument.

...
additional arguments passed to load_cytoset_from_fcs.

Value

object of class GatingSet.

Details

Calls are then made to cyto_markers_edit and cyto_details_edit to update the GatingSet with the details of the experiment. These details can be modified later with additional calls to cyto_markers_edit and/or cyto_details_edit.

Through the clean argument, the data can then be optionally cleaned using flow_auto_qc to automatically remove anomalies in the recorded data.

Users can optionally provide a name for a gatingTemplate csv file which will be created if necessary and assigned as the active gatingTemplate.

See also

Examples

if (FALSE) {
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Get path to Activation .fcs files in CytoExploreRData
datadir <- system.file("extdata", package = "CytoExploreRData")
path <- paste0(datadir, "/Activation")

# Load in .fcs files into an annotated GatingSet
gs <- cyto_setup(path)

# Markers have been assigned
cyto_extract(gs, "root")[[1]]

# Experiment details have been updated
cyto_details(gs)
}